backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
scikit-learn==1.7.1
joblib==1.5.1

